function/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-storage==1.31.0
SQLAlchemy==1.3.12
pg8000==1.13.2